renaming integrations to modules
