+++ automation/modules/base.py
@@ -1,6 +1,3 @@
-import imp
-
-
 import abc
 
 class Runnable( abc.ABC ) :
